Added: Feature/pr2 (#16)
## Description 

Please give a brief explanation of your changes. This part can be used
as message of the commit produced on `squash and merge`.

## Tests

Refer to executed tests in this section. Please link the runs of the
executed actions and give a description what exactly is proofed with the
test (IaC changes, workflow changes, etc.). This can also be a copy of
the output of a local run.

## Documentation

Add links for relevant documentation written for this change. The
documentation will also be part of this review.

## Additional Notes

Give additional comments here. Those could for example include:

* any manual actions that are required to merge this
* any issues that could occur when merging this PR
* special instructions for the release (e.g. alternative infrastructure
is required)
* related links to the change
* any changes (for example `experimental/*` branches or a test instance
config) which shall be deleted after the review

## Ticket

Add a link to the related ticket.

[<title>](<link>)

---------

Co-authored-by: lukas <lukas.schuster@quibiq.de>